feat : 로또 번호 유효성 검사 메서드 구현
